feat(archiver): Refactor archiver to use dynamic dispatch
This commit refactors the archiver module to use dynamic dispatch for different archive formats.
It introduces a new function `create_archiver` that returns a Boxed trait object of the Archiver trait based on the file extension of the destination path.
The `execute` method of the Archiver trait now takes ownership of the ArchiverOpts struct instead of a reference.
The `find_format` function is introduced in the format module to determine the archive format based on the file extension.
The main function is also updated to use the new archiver creation function.